requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.4
python-dotenv==1.0.1
google-genai==0.3.0
requests==2.32.3
nltk==3.8.1
pydantic==2.7.4
